Move ConfigureAssignmentSchema with other lti schemas
