+++ lms/validation/__init__.py
@@ -50,7 +50,6 @@ from lms.validation._api import (
     APIRecordResultSchema,
     APIRecordSpeedgraderSchema,
 )
-from lms.validation._assignment import ConfigureAssignmentSchema
 from lms.validation._base import (
     PlainSchema,
     PyramidRequestSchema,
@@ -59,6 +58,7 @@ from lms.validation._base import (
 from lms.validation._exceptions import LTIToolRedirect, ValidationError
 from lms.validation._lti_launch_params import (
     BasicLTILaunchSchema,
+    ConfigureAssignmentSchema,
     ContentItemSelectionLTILaunchSchema,
     LTIV11CoreSchema,
     URLConfiguredBasicLTILaunchSchema,
